+++ examples/digital_key_device_cs/digital_key_device_cs.c
@@ -1339,9 +1339,9 @@ static void BleApp_CsEventHandler(deviceId_t deviceId, void *pData, appCsEventTy
             if (pEvent->eventType == commandError_c)
             {
                 shell_write("CS Command Complete error! errorSource: ");
-                shell_writeDec(pEvent->eventData.csCommandError.errorSource); /* value in commandErrorSource_t enum */
+                shell_writeDec((uint32_t)pEvent->eventData.csCommandError.errorSource); /* value in commandErrorSource_t enum */
                 shell_write(", status ");
-                shell_writeDec(pEvent->eventData.csCommandError.status); /* value in bleResult_t enum */
+                shell_writeDec((uint32_t)pEvent->eventData.csCommandError.status); /* value in bleResult_t enum */
                 SHELL_NEWLINE();
             }
         }
